+++ HISTORY.md
@@ -9,6 +9,8 @@ V2.3.0 ()
 * Fixed ubus upgrade command cannot upgrade skynet version
 * API_VER 15 added app_name method in device object
 * Traverse devices' data when app watches input
+* Added snapshot in cov class
+* When ttl is less than 1 second then disable
 
 
 V2.2.4 (2023-10-11)